GH #755 - Working on AMQP.

+++ code/zato-server/src/zato/server/base/config_manager/amqp_.py
@@ -7,6 +7,7 @@ Licensed under AGPLv3, see LICENSE.txt for terms and conditions.
 """
 
 # stdlib
+from copy import deepcopy
 from logging import getLogger
 from traceback import format_exc
 
@@ -38,8 +39,13 @@ class AMQP(ConfigManagerImpl):
         msg:'Bunch',
     ) -> 'None':
         msg.password = self.server.decrypt(msg.password)
-        msg.is_active = True
-        self.amqp_api.create(msg.name, msg, self.invoke, needs_start=True)
+
+        # The connector is always active, only the channel or outconn under it can be inactive,
+        # so the connector gets its own copy of the config and the object's flag stays untouched.
+        connector_config = deepcopy(msg)
+        connector_config.is_active = True
+
+        self.amqp_api.create(msg.name, connector_config, self.invoke, needs_start=True)
 
 # ################################################################################################################################
 
@@ -58,11 +64,21 @@ class AMQP(ConfigManagerImpl):
         self:'ConfigManager', # type: ignore
         msg:'Bunch',
     ) -> 'None':
+        """ Replaces the connector of an outgoing connection with one built from the new configuration,
+        which is what makes a new name, address or credentials take effect at runtime.
+        """
         msg.password = self.server.decrypt(msg.password)
         with self.update_lock:
             del self.amqp_out_name_to_def[msg.old_name]
             self.amqp_out_name_to_def[msg.name] = msg.name
-            self.amqp_api.edit_outconn(msg.name, msg)
+
+            # The connector may be absent if it could not be created at startup ..
+            if msg.old_name in self.amqp_api.connectors:
+                _ = self.amqp_api.delete(msg.old_name)
+
+            # .. and the new one is built exactly the way a freshly created connection is.
+            self.amqp_connection_create(msg)
+            self.amqp_api.create_outconn(msg.name, msg)
 
 # ################################################################################################################################
 
@@ -70,9 +86,15 @@ class AMQP(ConfigManagerImpl):
         self:'ConfigManager', # type: ignore
         msg:'Bunch',
     ) -> 'None':
+        """ Removes an outgoing connection along with its connector.
+        """
         with self.update_lock:
             del self.amqp_out_name_to_def[msg.name]
-            self.amqp_api.delete_outconn(msg.name, msg)
+
+            # The connector may be absent if it could not be created at startup.
+            if msg.name in self.amqp_api.connectors:
+                self.amqp_api.delete_outconn(msg.name, msg)
+                _ = self.amqp_api.delete(msg.name)
 
 # ################################################################################################################################
 
@@ -90,9 +112,19 @@ class AMQP(ConfigManagerImpl):
         self:'ConfigManager', # type: ignore
         msg:'Bunch',
     ) -> 'None':
+        """ Replaces the connector of a channel with one built from the new configuration,
+        which is what makes a new name, address or credentials take effect at runtime.
+        """
         msg.password = self.server.decrypt(msg.password)
         with self.update_lock:
-            self.amqp_api.edit_channel(msg.name, msg)
+
+            # The connector may be absent if it could not be created at startup ..
+            if msg.old_name in self.amqp_api.connectors:
+                _ = self.amqp_api.delete(msg.old_name)
+
+            # .. and the new one is built exactly the way a freshly created channel is.
+            self.amqp_connection_create(msg)
+            self.amqp_api.create_channel(msg.name, msg)
 
 # ################################################################################################################################
 
@@ -100,8 +132,14 @@ class AMQP(ConfigManagerImpl):
         self:'ConfigManager', # type: ignore
         msg:'Bunch',
     ) -> 'None':
+        """ Removes a channel along with its connector.
+        """
         with self.update_lock:
-            self.amqp_api.delete_channel(msg.name, msg)
+
+            # The connector may be absent if it could not be created at startup.
+            if msg.name in self.amqp_api.connectors:
+                self.amqp_api.delete_channel(msg.name, msg)
+                _ = self.amqp_api.delete(msg.name)
 
 # ################################################################################################################################
 
@@ -136,7 +174,7 @@ class AMQP(ConfigManagerImpl):
         *args:'any_',
         **kwargs:'any_',
     ) -> 'None':
-        spawn_greenlet(self._amqp_invoke_async, *args, **kwargs)
+        _ = spawn_greenlet(self._amqp_invoke_async, *args, **kwargs)
 
 # ################################################################################################################################
 # ################################################################################################################################
